AE-1772: moving classes from inventory-plugin mojo to library
- ExecutionStatus
- ExecutionStatusEntry

and extending signature to include artifact
